Added automatic localization addition
Settings now has a PoolStringArray of locales (in place of the constants) and a equally indexed PoolStringArray of language names.
Localization files now are required to provide a name via the LANGUAGE_NAME key which must be in the first line of the file.
SettingsDialog now automatically sets the items in the OptionButton.