Sorted languages alphabetically, with English on top
Separated the adding of items to its own function. which adds them in alphabetical order, storing the original position in the id.
Also changed the callback method for the OptionButton to use id and not index to set the locale.
And initialized locales and language_names on declaration to clear warnings.