Retranslate pt_BR (#192)
A retranslation of `pt_BR` to improve the fidelity (to `en`),
consistency, and avoid confusion and ambiguity. The previously existing
translation was taken into consideration during retranslation.

"Consistency" as mentioned means e.g. always translating a word to the
same term when it's used with the same meaning, and following `en`'s
patterns (window and dialogue titles are capitalized only in the first
word and more descriptive tools/option names as well).